requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
motor==3.1.2
python-dotenv==1.0.0
pydantic==1.10.7